+++ src/settings/AudioSettings.tsx
@@ -6,6 +6,11 @@ import Select from '@mui/material/Select';
 import Stack from '@mui/material/Stack';
 import { ObsAudioDevice } from 'main/obsAudioDeviceUtils';
 import ConfigContext from "./ConfigContext";
+import Box from '@mui/material/Box';
+import InfoIcon from '@mui/icons-material/Info';
+import IconButton from '@mui/material/IconButton';
+import Tooltip from '@mui/material/Tooltip';
+import { configSchema } from '../main/configSchema'
 
 const ipc = window.electron.ipcRenderer;
 
@@ -34,7 +39,7 @@ export default function GeneralSettings() {
   };
 
   const style = {
-    height: '2.5rem',
+    width: '405px',
     color: 'white',
     '& .MuiOutlinedInput-notchedOutline': {
       borderColor: 'black'
@@ -52,45 +57,59 @@ export default function GeneralSettings() {
     <Stack
       component="form"
       sx={{
-        '& > :not(style)': { m: 1, width: '50ch' },
+        '& > :not(style)': { m: 0, width: '50ch' },
       }}
       noValidate
       autoComplete="off"
     >
-      <FormControl fullWidth>
-        <InputLabel id="demo-simple-select-label" sx = {style}>Input</InputLabel>
-        <Select
-          labelId="demo-simple-select-label"
-          id="audio-input-device"
-          value={config.audioInputDevice}
-          label="Input"
-          onChange={(event) => modifyConfig('audioInputDevice', event.target.value)}
-          sx={style}
-        >
-          { availableAudioDevices.input.map((device: ObsAudioDevice) => {
-            return (
-              <MenuItem key={ 'device_' + device.id } value={ device.id }>{ device.name }</MenuItem>
-            )
-          })}
-        </Select>
-      </FormControl>
-      <FormControl fullWidth>
-        <InputLabel id="demo-simple-select-label" sx = {style}>Output</InputLabel>
-        <Select
-          labelId="demo-simple-select-label"
-          id="audio-output-device"
-          value={config.audioOutputDevice}
-          label="Output"
-          onChange={(event) => modifyConfig('audioOutputDevice', event.target.value)}
-          sx={style}
-        >
-          { availableAudioDevices.output.map((device: ObsAudioDevice) => {
-            return (
-              <MenuItem key={ 'device_' + device.id } value={ device.id }>{ device.name }</MenuItem>
-            )
-          })}
-        </Select>
-      </FormControl>
+      <Box component="span" sx={{ display: 'flex', alignItems: 'center'}}>
+        <FormControl sx={{my: 1}}>
+          <InputLabel id="demo-simple-select-label" sx = {style}>Input</InputLabel>
+          <Select
+            labelId="demo-simple-select-label"
+            id="audio-input-device"
+            value={config.audioInputDevice}
+            label="Input"
+            onChange={(event) => modifyConfig('audioInputDevice', event.target.value)}
+            sx={style}
+          >
+            { availableAudioDevices.input.map((device: ObsAudioDevice) => {
+              return (
+                <MenuItem key={ 'device_' + device.id } value={ device.id }>{ device.name }</MenuItem>
+              )
+            })}
+          </Select>
+        </FormControl>
+        <Tooltip title={configSchema["audioInputDevice"].description} >
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
+      <Box component="span" sx={{ display: 'flex', alignItems: 'center'}}>
+        <FormControl sx={{my: 1}}>
+          <InputLabel id="demo-simple-select-label" sx = {style}>Output</InputLabel>
+          <Select
+            labelId="demo-simple-select-label"
+            id="audio-output-device"
+            value={config.audioOutputDevice}
+            label="Output"
+            onChange={(event) => modifyConfig('audioOutputDevice', event.target.value)}
+            sx={style}
+          >
+            { availableAudioDevices.output.map((device: ObsAudioDevice) => {
+              return (
+                <MenuItem key={ 'device_' + device.id } value={ device.id }>{ device.name }</MenuItem>
+              )
+            })}
+          </Select>
+        </FormControl>
+        <Tooltip title={configSchema["audioOutputDevice"].description} >
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
     </Stack>
   );
 }