+++ src/settings/GeneralSettings.tsx
@@ -6,6 +6,11 @@ import Checkbox from '@mui/material/Checkbox';
 import FormGroup from '@mui/material/FormGroup';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import ConfigContext from "./ConfigContext";
+import InfoIcon from '@mui/icons-material/Info';
+import IconButton from '@mui/material/IconButton';
+import Tooltip from '@mui/material/Tooltip';
+import Box from '@mui/material/Box';
+import { configSchema } from '../main/configSchema'
 
 const ipc = window.electron.ipcRenderer;
 
@@ -34,6 +39,7 @@ export default function GeneralSettings() {
   }, []);
 
   const style = {
+    width: '450px',
     "& .MuiOutlinedInput-root": {
       "&.Mui-focused fieldset": {borderColor: "#bb4220"},
       "& > fieldset": {borderColor: "black" }
@@ -61,31 +67,46 @@ export default function GeneralSettings() {
     <Stack
       component="form"
       sx={{
-        '& > :not(style)': { m: 1, width: '50ch' },
+        '& > :not(style)': { m: 0, width: '50ch' },
       }}
       noValidate
       autoComplete="off"
     >
-      <TextField 
-        value={config.storagePath}
-        id="storage-path" 
-        label="Storage Path" 
-        variant="outlined" 
-        onClick={() => openDirectorySelectorDialog("storagePath")} 
-        InputLabelProps={{ shrink: true }}
-        sx={style}
-        inputProps={{ style: { color: "white" } }}
-      />
-      <TextField 
-        value={config.retailLogPath}
-        id="retail-log-path" 
-        label="Retail Log Path" 
-        variant="outlined" 
-        onClick={() => openDirectorySelectorDialog("retailLogPath")} 
-        InputLabelProps={{ shrink: true }}
-        sx={style}
-        inputProps={{ style: { color: "white" } }}
-      />
+      <Box component="span" sx={{ display: 'flex', alignItems: 'flex-start' }}>
+        <TextField 
+          value={config.storagePath}
+          id="storage-path" 
+          label="Storage Path" 
+          variant="outlined" 
+          onClick={() => openDirectorySelectorDialog("storagePath")} 
+          InputLabelProps={{ shrink: true }}
+          sx={{...style, my: 1}}
+          inputProps={{ style: { color: "white" } }}
+        />
+        <Tooltip title={configSchema["storagePath"].description} sx={{position: 'relative', right: '0px', top: '17px'}}>
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
+      
+      <Box component="span" sx={{ display: 'flex', alignItems: 'flex-start' }}>
+        <TextField 
+          value={config.retailLogPath}
+          id="retail-log-path" 
+          label="Retail Log Path" 
+          variant="outlined" 
+          onClick={() => openDirectorySelectorDialog("retailLogPath")} 
+          InputLabelProps={{ shrink: true }}
+          sx={{...style, my: 1}}
+          inputProps={{ style: { color: "white" } }}
+        />
+        <Tooltip title={configSchema["retailLogPath"].description} sx={{position: 'relative', right: '0px', top: '17px'}}>
+          <IconButton>
+            <InfoIcon style={{ color: 'white'}}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
       {/* <TextField 
         value={config.classicLogPath}
         id="classic-log-path" 
@@ -96,22 +117,50 @@ export default function GeneralSettings() {
         sx={style}
         inputProps={{ style: { color: "white" } }}
       /> */}
-      <TextField 
-        value={config.maxStorage}
-        onChange={event => { modifyConfig("maxStorage", parseInt(event.target.value, 10)) }}
-        id="max-storage" 
-        label="Max Storage (GB)" 
-        variant="outlined" 
-        type="number" 
-        error= { config.maxStorage < 0 }
-        helperText={(config.maxStorage < 0) ? "Must be positive" : ' '}
-        InputLabelProps={{ shrink: true }}
-        sx={style}
-        inputProps={{ style: { color: "white" } }}
-      />
-      <FormGroup sx={formGroupStyle}>
-        <FormControlLabel control={getCheckBox("startUp")} label="Run on startup" style = {formControlLabelStyle} />
-      </FormGroup>
+
+      <Box component="span" sx={{ display: 'flex', alignItems: 'flex-start' }}>
+        <TextField 
+          value={config.maxStorage}
+          onChange={event => { modifyConfig("maxStorage", parseInt(event.target.value, 10)) }}
+          id="max-storage" 
+          label="Max Storage (GB)" 
+          variant="outlined" 
+          type="number" 
+          error= { config.maxStorage < 0 }
+          helperText={(config.maxStorage < 0) ? "Must be positive" : ' '}
+          InputLabelProps={{ shrink: true }}
+          sx={{...style, my: 1}}
+          inputProps={{ style: { color: "white" } }}
+        />
+        <Tooltip title={configSchema["maxStorage"].description} sx={{position: 'relative', right: '0px', top: '17px'}}>
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
+
+      <Box component="span">
+        <FormGroup sx={formGroupStyle}>
+          <FormControlLabel control={getCheckBox("startUp")} label="Run on startup" style = {formControlLabelStyle} />
+        </FormGroup>
+        <Tooltip title={configSchema["startUp"].description} sx={{position: 'fixed', left: '315px', top: '292px'}}>
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
+
+      <Box component="span">
+        <FormGroup sx={formGroupStyle}>
+          <FormControlLabel control={getCheckBox("startMinimized")} label="Start minimized" style={formControlLabelStyle} />
+        </FormGroup>
+        <Tooltip title={configSchema["startMinimized"].description} sx={{position: 'fixed', left: '315px', top: '333px'}}>
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
+      
     </Stack>
   );
 }